refactor: add env vars for enabling features like copilot or work stuff

+++ .dotfiles/shell/aliases.sh
@@ -305,26 +305,28 @@ if is-supported docker; then
     ## - - - - - - - - - - - - - - - - - - - - - - - - - - -  }}}
 
     ## work docker aliases for calcite-components             {{{
-    # workaround to run e2e tests on ubuntu due to a Stencil bug
-    # https://github.com/ionic-team/stencil/issues/3853
-    # Uses a bind mount so works for local development
-    cc_docker_cmd="docker run --init --interactive --rm --cap-add SYS_ADMIN --volume .:/app:z --user $(id -u):$(id -g)"
+    if [ "$USE_WORK_STUFF" = "1" ]; then
+        # workaround to run e2e tests on ubuntu due to a Stencil bug
+        # https://github.com/ionic-team/stencil/issues/3853
+        # Uses a bind mount so works for local development
+        cc_docker_cmd="docker run --init --interactive --rm --cap-add SYS_ADMIN --volume .:/app:z --user $(id -u):$(id -g)"
 
-    # shellcheck disable=2139
-    alias cc_start_in_docker="$cc_docker_cmd --publish 3333:3333 --name calcite-components-start calcite-components npm --workspace=@esri/calcite-components start"
+        # shellcheck disable=2139
+        alias cc_start_in_docker="$cc_docker_cmd --publish 3333:3333 --name calcite-components-start calcite-components npm --workspace=@esri/calcite-components start"
 
-    # shellcheck disable=2139
-    alias cc_test_in_docker="$cc_docker_cmd --name calcite-components_test calcite-components npm --workspace=@esri/calcite-components run test -- -- --watch"
+        # shellcheck disable=2139
+        alias cc_test_in_docker="$cc_docker_cmd --name calcite-components_test calcite-components npm --workspace=@esri/calcite-components run test -- -- --watch"
 
-    # shellcheck disable=2139
-    alias cc_run_in_docker="$cc_docker_cmd --name calcite-components_run calcite-components npm --workspace=@esri/calcite-components run"
+        # shellcheck disable=2139
+        alias cc_run_in_docker="$cc_docker_cmd --name calcite-components_run calcite-components npm --workspace=@esri/calcite-components run"
 
-    alias cc_build_docker_image="docker build --tag calcite-components ."
+        alias cc_build_docker_image="docker build --tag calcite-components ."
 
-    # I need to link the Dockerfile to the current git worktree
-    alias cc_link_dockerfile='ln -f $WORK/calcite-design-system/Dockerfile'
+        # I need to link the Dockerfile to the current git worktree
+        alias cc_link_dockerfile='ln -f $WORK/calcite-design-system/Dockerfile'
 
-    unset cc_docker_cmd
+        unset cc_docker_cmd
+    fi
 
     ## - - - - - - - - - - - - - - - - - - - - - - - - - - -  }}}
 fi
